move over last constraints from v1
